// web/generated/agentvouch/src/generated/types/authorDisputeLiabilityScope.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum AuthorDisputeLiabilityScope {
  AuthorBondOnly,
  AuthorBondThenVouchers,
}

export type AuthorDisputeLiabilityScopeArgs = AuthorDisputeLiabilityScope;

export function getAuthorDisputeLiabilityScopeEncoder(): FixedSizeEncoder<AuthorDisputeLiabilityScopeArgs> {
  return getEnumEncoder(AuthorDisputeLiabilityScope);
}

export function getAuthorDisputeLiabilityScopeDecoder(): FixedSizeDecoder<AuthorDisputeLiabilityScope> {
  return getEnumDecoder(AuthorDisputeLiabilityScope);
}

export function getAuthorDisputeLiabilityScopeCodec(): FixedSizeCodec<
  AuthorDisputeLiabilityScopeArgs,
  AuthorDisputeLiabilityScope
> {
  return combineCodec(
    getAuthorDisputeLiabilityScopeEncoder(),
    getAuthorDisputeLiabilityScopeDecoder(),
  );
}