// web/generated/agentvouch/src/generated/types/authorDisputeReason.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum AuthorDisputeReason {
  MaliciousSkill,
  FraudulentClaims,
  FailedDelivery,
  Other,
}

export type AuthorDisputeReasonArgs = AuthorDisputeReason;

export function getAuthorDisputeReasonEncoder(): FixedSizeEncoder<AuthorDisputeReasonArgs> {
  return getEnumEncoder(AuthorDisputeReason);
}

export function getAuthorDisputeReasonDecoder(): FixedSizeDecoder<AuthorDisputeReason> {
  return getEnumDecoder(AuthorDisputeReason);
}

export function getAuthorDisputeReasonCodec(): FixedSizeCodec<
  AuthorDisputeReasonArgs,
  AuthorDisputeReason
> {
  return combineCodec(
    getAuthorDisputeReasonEncoder(),
    getAuthorDisputeReasonDecoder(),
  );
}