// web/generated/agentvouch/src/generated/types/authorDisputeRuling.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum AuthorDisputeRuling {
  Upheld,
  Dismissed,
}

export type AuthorDisputeRulingArgs = AuthorDisputeRuling;

export function getAuthorDisputeRulingEncoder(): FixedSizeEncoder<AuthorDisputeRulingArgs> {
  return getEnumEncoder(AuthorDisputeRuling);
}

export function getAuthorDisputeRulingDecoder(): FixedSizeDecoder<AuthorDisputeRuling> {
  return getEnumDecoder(AuthorDisputeRuling);
}

export function getAuthorDisputeRulingCodec(): FixedSizeCodec<
  AuthorDisputeRulingArgs,
  AuthorDisputeRuling
> {
  return combineCodec(
    getAuthorDisputeRulingEncoder(),
    getAuthorDisputeRulingDecoder(),
  );
}